feat: 유저 삭제 시 에러 로직 추가

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -111,6 +111,11 @@ public class BasicUserService implements UserService {
     @Override
     public void delete(UUID userId) {
         User user = userRepository.findById(userId);
+
+        if (user == null) {
+            throw new UserNotFoundException(userId + " 유저를 찾을 수 없습니다.");
+        }
+
         userRepository.delete(user.getId());
 
         UserStatus userStatus = userStatusService.findById(user.getId());
